fix: workaround for Object::GetIsolate
- Object::GetIsolate
It will return the current isolate always rather than the real isolate
the object belongs to. This is because there is no Escargot api as of
now to acquire a VMInstance by an esObject pointer.

- FunctionTemplate::New
We don't support SideEffect using JS-observable that seems to be mainly
used by v8-inspector.

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com