refactor: BackingStoreWrap
We need to support two types of BackingStore. One is to use the array
buffer allocator of the isolate. The other is to take over the ownership
of the user's given buffer. The latter requires handling a deleter
callback which should be invoked at the time the destructor called.

For supporting that, I suggest refactoring the current BackingStoreWrap.
BackingStoreWrap is related to BufferDeleter instead of
v8::ArrayBuffer::Allocator.

Resolve #545

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com