fix: workaround for Escargot missing feature (look-behind assertion)
This workaround is for `look-behind assertion` issued: #542.

I just removed the unsupported part from the expression for now.
It seems not critical. The feature's used during parsing a formatted
string for an error call stacks message. Thus, we may fix this line
again if the error message isn't printed properly.

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com