feat: workaround for Object::CreationContext
In order to save memory, this commit uses only the current context
for Object::CreationContext.

Reason:
In Node.js at this time, one main Context associated with the
Environment instance is used for most Node.js features (except writing
MessagePort objects.) So, on purpose, we don't store Object's creation
context which is related to Object::CreateContext().

Observation:
As of now, one cctest case is found. Two contexts are used but not
related to this issue.

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com