fix: disable the assertion for flust_task_
It seems an implementation coupled with V8 engine behaviour.

When terminated through DefaultProcessExitHandler, the assertion occurs
in ~PerIsolatePlatformData. With Node.js with v8, the shared instance of
PerIsolatePlatformData is being used at the termination. So the
assertion in ~PerIsolatePlatformData doesn't effect the termination.

However lwnode doesn't use PerIsolatePlatformData at that time. So the
destructor is invoked.

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com