feat: add HandleScopeWrapGuard
This is for the case where calling JS function in memory-safe wise.
In V8, it should be handled-scoped to call JS function from Native
function in order to preventing memory leak. If we want to call such
a function only using Escargot API, there is no tool to scope the
space for handles that may be allocated during calling js function.

example:
```c++
{
  HandleScopeWrapGuard guard(IsolateWrap::GetCurrent());
  ...
  Evaluator::execute(
      esContext,
      [](ExecutionStateRef* state,
         ValueRef* self,
         FunctionObjectRef* function,
         size_t argc,
         ValueRef** argv) -> ValueRef* {
        // this call may be associated to HandleWrap allocation
        // if a HandleScope used in V8 is sealed, then it will
        // be handled as error.
        return function->call(state, self, argc, argv);
      },
      target,
      esFunction,
      arguments.size(),
      arguments.data());
}
```

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com