chore: run tests on the node.js root
changes the cwd to `deps/lwnode` before running the Node.js tests.
It will ensure all the tests to run as intended.

I found the two followings:

- Some Node.js tests have been written upon the assumtion that the
current working directory is always the Node.js root, `deps/lwnode`.

- `--test-root` option of `lwnode/test.py`, which we uses to run the
test on the root of the shim directories, seems not officially managed.

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com